% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_gama.R
\name{call_gama}
\alias{call_gama}
\title{Run GAMA on a XML file}
\usage{
call_gama(parameter_xml_file, hpc, output_dir = "")
}
\arguments{
\item{parameter_xml_file}{path to an XML file containing an experiment.}

\item{hpc}{numeric, number of cores}

\item{output_dir}{path to saved the output of gama. If not specified, current
working directory will be used. If `output_dir` doesn't
exist, it will be created.}
}
\description{
From a XML file containing an experiment plan, send and run it in GAMA and
returns one XML file by simulation containing the output.
}
\examples{
parameter_xml_file <- system.file("models", "sir.xml", package = "gamar")
outfiles <- call_gama(parameter_xml_file, hpc = 1, output_dir = tempdir())
}
