% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{defpath}
\alias{defpath}
\title{Define the GAMA path}
\usage{
defpath(path)
}
\arguments{
\item{path}{Absolute path to the folder that contains the GAMA Java plugins.}
}
\description{
Define the path to GAMA Java plugins.
}
