% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examples.R
\name{examples}
\alias{examples}
\title{Manage built-in models}
\usage{
examples(what, to)
}
\arguments{
\item{what}{character string, name of the built-in model to copy, as defined
when calling the function without arguments.}

\item{to}{character string, path of the location to copy to.}
}
\description{
This function can be used with 0, 1 or 2 arguments values. With no argument
values it lists the built-in models available in the system library. With 1
or 2 argument values it will copy (a) built-in model(s) to (a) specified
location(s).
}
