% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R, R/experiment.R, R/is.experiment.R
\name{[.experiment}
\alias{[.experiment}
\alias{experiment}
\alias{is.experiment}
\title{Extract or Replace Parts of an Experiment Object}
\usage{
\method{[}{experiment}(exp, i, j, ..., drop = TRUE)

experiment(parameters, obsrates, tmax, seed, experiment, model,
  dic = NULL)

is.experiment(exp)
}
\arguments{
\item{exp}{object to be tested}

\item{i, j, ...}{indices specifying elements to extract or replace. Indices are
numeric or character vectors or empty (missing) or NULL.}

\item{drop}{boolean, TRUE the result is coerced to the lowest possible
dimension.}

\item{parameters}{A data frame of numerical values giving the values of each
parameter (in column), for each simulation (in row).}

\item{obsrates}{A data frame of positive integer values giving the periods,
in time steps, at which observed variables are observed.
Should have the same number of rows as \code{parameters}.}

\item{tmax}{A positive integer vector, the length of which is either 1 or
equal to the number of \code{parameters} and \code{obsrates}: it
gives the end of simulations, in numbers of time steps.}

\item{seed}{A numerical vector, the length of which is either 1 or equal to
the number of \code{parameters} and \code{obsrates}: it gives the
seeds of the simulations.}

\item{experiment}{The name of an experiment of the \code{GAML} file
\code{model}.}

\item{model}{The path to a \code{GAML} file.}

\item{dic}{A named vector of character strings. The values and the names of
this vector should be consistent with the names of
\code{parameters}, \code{obsrates} as well as the variables and
parameters defined in the \code{model} \code{GAML} file. See
Details for more information.}
}
\value{
An object of class \code{experiment}

An object of class \code{experiment}.

The function returns `TRUE` or `FALSE` depending on whether its
argument is of chatacter type or not
}
\description{
Extracts or replaces parts of an \code{experiment} object with new value(s).

Allows to build an object of class \code{experiment} from individual parts.
}
\details{
The class \code{experiment} inherits from the class \code{tibble}
(\code{tbl_df}). It contains parameters values as well as periods of
observation of the observed variables and it connects to a \code{GAML} file
that specifies the full model.

TO DO: Explains what \code{dic} is.
}
\examples{
## experiment ##
exp0 <- experiment(
  data.frame(S0 = 999, I0 = 1, R0 = 0, beta = 1.5, gamma = .15),
  data.frame(S = 1, I = 1, R = 1),
  1000, 1, "sir", system.file("models", "sir.gaml", package = "gamar"))
print(exp0)

## is.experiment ##
# to test if an object is a class `experiment`
df <- data.frame("S0" = rep(999, 5), "I0" = rep(1, 5), "R0" = rep(0, 5),
                 "beta" = rep(1.5, 5), "gamma" = runif (5, 0, 1),
                 "S" = rep(1, 5), "I" = rep(1, 5), "R" = rep(1, 5),
                 "a" = rep(1000, 5), "b" = rep(1, 5))
exp <- as_experiment(df, parameters = c("S0", "I0", "R0", "beta", "gamma"),
                     obsrates = c("S", "I", "R"),
                     tmax = "a",
                     seed = "b",
                     experiment = "sir",
                     model =
                       system.file("models", "sir.gaml", package = "gamar"))
is.experiment(exp)



# Here is an experiment with 1 simulation:
sir1 <- load_experiment("sir",
                        system.file("models", "sir.gaml", package = "gamar"))
sir1

sir2 <- sir1
# If  we now replace the values of "p_S0" of "sir2" by a single value:
sir2$p_S0 <- 2
sir2
# If you wish to delete one column:
sir2$r_R <- NULL
sir2
}
