% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fullfact.R
\name{fullfact}
\alias{fullfact}
\title{Populate an experiment with values}
\usage{
fullfact(xprmt, ...)
}
\arguments{
\item{xprmt}{An object of class \code{experiment}.}

\item{...}{Vectors used to overwrite columns of \code{xprmt}. These arguments
should be named and their names should match the names of
\code{xprmt}.}
}
\value{
An object of class \code{experiment} with a full factorial design of
the inputed \code{experiment}.
}
\description{
Generates a full experimental design for an experiment.
}
\details{
This function is basically a wrapper around the
\code{\link[base]{expand.grid}} function applied to all the column of the
inputed \code{experiment} object. Any of these columns can be replaced by
user-defined vectors of values, see examples.
}
\examples{
sir1 <- load_experiment("sir",
                        system.file("models", "sir.gaml", package = "gamar"))
# 1. First type of use: tranforming an experiment into one with a full
# factorial design: (use the function `repl`, if you want more details `?repl`)
sir2 <- repl(sir1, 3)
sir2$p_S0 <- 1:3
sir2
sir2[1, 2] <- 2
# "sir2" is not full factorial:
sir2
# this is:
fullfact(sir2)

# 2. Second type of use: by providing vectors of values to overwrite elements
# of the "experiment" object and then expand it into full factorial design:
fullfact(sir2, p_S0 = 1:3, p_I0 = 4:5)
}
