% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{is_gama_installed}
\alias{is_gama_installed}
\title{Test if GAMA is intalled}
\usage{
is_gama_installed(path = unlist(options("gamar.gama.path")))
}
\arguments{
\item{path}{path to the application Gama platform}
}
\description{
Test if GAMA is intalled and is correctly linked to the application Gama
platform
}
\examples{
\dontrun{
# For examples, for MacOs, it can be:
is_gama_installed(path = "/Applications/Gama.app")
}
}
