% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_experiments.R
\name{list_experiments}
\alias{list_experiments}
\title{List the experiments of a model and their types}
\usage{
list_experiments(file)
}
\arguments{
\item{file}{Path to a \code{.gaml} file.}
}
\description{
List the experiments of a given model.
}
\examples{
gaml_file <- system.file("models", "sir.gaml", package = "gamar")
list_experiments(gaml_file)
}
