% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment.R
\name{load_experiment}
\alias{load_experiment}
\title{Loads An Experiment}
\usage{
load_experiment(exprmnt, model)
}
\arguments{
\item{exprmnt}{a character string giving the name of the experiment to load}

\item{model}{a character string giving the name of the file from which to
load the experiment.}
}
\description{
Loads an experiment from a model specified in a gaml file and returns an
object of class \code{plan}.
}
