% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_movie.R
\name{make_movie}
\alias{make_movie}
\title{Create GIF animations}
\usage{
make_movie(exp, name_col, output_path = NULL, autoplay = TRUE,
  interval = 0.01, loop = TRUE, autobrowse = FALSE, ...)
}
\arguments{
\item{exp}{an object of class experiment.}

\item{name_col}{character string, name of the column containing the display
information (format .png).}

\item{output_path}{path to store the .gif file generated, if \code{NULL} will
take the path store in the attributes of each data frame contained in the
column \code{output}.}

\item{autoplay}{logical, whether to autoplay the animation when the HTML
page is loaded (default to be TRUE).}

\item{interval}{a positive number to set the time interval of the animation
(unit in seconds); default to be 0.01.}

\item{loop}{logical or numeric, Number of times the GIF animation is to
cycle; default TRUE (infinite loop)}

\item{autobrowse}{logical: whether auto-browse the animation page immediately
after it is created?; default to be FALSE}

\item{...}{arguments in tag = value form, or a list of tagged values.
The tags usually come from the animation parameters.}
}
\value{
path the gif file(s) invisibly.
}
\description{
From the output of the function \code{run_experiment} with display, create
a .gif file of the output for one specific columns. One gif will be created
for each experiment (rows in the input object).
}
\details{
The function uses the package \code{animation} to generate the .gif file(s).
Please take a look at the documentation of the package and to the website
\url{ https://yihui.name/animation} for more information on the usage and
the installation of this package.\cr\cr
The arguments \code{autoplay}, \code{interval}, \code{loop},
\code{autobrowse} and \code{...} are used to control the behaviour of the
animation, see \code{?animation::ani.options} for more details..
}
\examples{
# Load an experiment
exp <- load_experiment("prey_predator",
                       system.file("models/predator_prey/models",
                                   "predator_prey.gaml", package = "gamar"))

# to explore the parameters and the observed variables
parameters(exp)
obs_rates(exp)

# to set the experiment
exp$p_Initial_number_of_preys_ <- 990L
exp$p_Initial_number_of_predators_ <- 100L
exp$p_Predator_probability_reproduce_ <- 0.1
exp$tmax <- 100L

# to run the experiment
out <- run_experiment(exp, hpc = 2, display = TRUE)

# to visualize the output
with(out$output[[1]],
    plot(Step, `r_Number_of_preys`, type = "l", lwd = 2, col = "red"))

# to make a movie
path_movie <- make_movie(out, "r_main_display")

# to print the result, it is necessary to have the package `magick`
magick::image_read(path_movie)
}
