% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_experiment.R
\name{map_experiment}
\alias{map_experiment}
\title{Map a data frame on an experiment}
\usage{
map_experiment(df, exp)
}
\arguments{
\item{df}{A data frame containing variables and parameters values for a
number of simulations we wish to use for the experiment \code{exp}.}

\item{exp}{An object of class \code{experiment}. See \code{Details} for
constraints on this object.}
}
\description{
Maps the data of a data frame on the structure on an object of class
\code{experiment}.
}
\details{
The names of the columns of the data frame should correspond to the names of
the variables and parameters of the experiment. This can work only if one of
the following conditions is met:
\itemize{
  \item the sets of names of the data frame and experiment are identical;
  \item the names of the data frame are included in the names of the
        experiment AND the numbers of rows of the data frame and the
        experiment are identical.
}
}
\examples{
# First situation: the sets of names of the data frame and the experiment
# are exactly the same:
if (exists("sir1", inherits = FALSE)) rm(sir1)
exp <- sir1
df <- as.data.frame(repl(sir1, 3))
map_experiment(df, exp)
# Second situation: the names of the data frame are included in the names of
# the experiment AND the numbers of rows of the data frame and the experiment
# are equal:
exp <- repl(sir1, 3)
df <- as.data.frame(exp)[, c(1, 3, 6)]
map_experiment(df, exp)
}
