% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{model}
\alias{model}
\alias{model.default}
\alias{model.experiment}
\alias{model<-}
\alias{model<-.default}
\alias{model<-.experiment}
\title{Get/Set the model of an experiment}
\usage{
model(exp)

\method{model}{default}(exp)

\method{model}{experiment}(exp)

model(exp) <- value

\method{model}{default}(exp) <- value

\method{model}{experiment}(exp) <- value
}
\arguments{
\item{exp}{An object of class \code{experiment}.}

\item{value}{Path of new gaml model file}
}
\value{
The path to the \code{.gaml} file that contains the model definition.
}
\description{
These functions allow to get and set the path to the \code{.gaml} file that
contains the model of an \code{experiment} object.
}
\examples{
exp1 <- load_experiment("sir", system.file("models", "sir.gaml",
                        package = "gamar"))
model(exp1)
model(exp1) <-  system.file("models", "CopyOfsir.gaml",
                        package = "gamar")
model(exp1)
exp1 <- load_experiment("sir", system.file("models", "sir.gaml",
                        package = "gamar"))
model(exp1)
model(exp1) <-  system.file("models", "CopyOfsir.gaml",
                        package = "gamar")
model(exp1)
}
