% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/name.R
\name{name}
\alias{name}
\alias{name.default}
\alias{name.experiment}
\title{Get the name of an experiment}
\usage{
name(x)

\method{name}{default}(x)

\method{name}{experiment}(x)
}
\arguments{
\item{x}{An object of class \code{experiment}.}
}
\value{
The name of the experiment that the inputed \code{experiment} object
is linked to.
}
\description{
Retrieves the name of the experiment that an \code{experiment} object is
linked to.
}
\examples{
exp1 <- load_experiment("sir",
                        system.file("models", "sir.gaml", package = "gamar"))
name(exp1)
}
