% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obs_rates.R
\name{obs_rates}
\alias{obs_rates}
\alias{obs_rates.default}
\alias{obs_rates.experiment}
\title{Extract monitoring rates}
\usage{
obs_rates(exp)

\method{obs_rates}{default}(exp)

\method{obs_rates}{experiment}(exp)
}
\arguments{
\item{exp}{An object of class \code{experiment}.}
}
\value{
A data frame that is a subset of the inputed \code{experiment} object.
}
\description{
Subsets the columns of an \code{experiment} object that correspond to the
obs_rates rates of the monitored variables.
}
\examples{
exp1 <- load_experiment("sir",
                        system.file("models", "sir.gaml", package = "gamar"))
exp2 <- repl(exp1, 10)
obs_rates(exp2)
}
