% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{parameters}
\alias{parameters}
\alias{parameters.default}
\alias{parameters.experiment}
\title{Extract parameters values}
\usage{
parameters(exp)

\method{parameters}{default}(exp)

\method{parameters}{experiment}(exp)
}
\arguments{
\item{exp}{An object of class \code{experiment}.}
}
\value{
A data frame that is a subset of the inputed \code{experiment} object.
}
\description{
Subsets the columns of an \code{experiment} object that correspond to the
values of the parameters.
}
\examples{
exp1 <- load_experiment("sir",
                        system.file("models", "sir.gaml", package = "gamar"))
exp2 <- repl(exp1, 10)
parameters(exp2)
}
