% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_params.R
\name{plot_params}
\alias{plot_params}
\title{Visualize values of parameters}
\usage{
plot_params(exp)
}
\arguments{
\item{exp}{An object of class \code{experiment}.}
}
\value{
Returns a vector of the variables with highest variances.
}
\description{
Plot the parameters that have the greatest variance in the experiments of a
model.
}
\details{
From the list of all parameters of the experiments that do vary (i.e. not
null) the one (up to 3) with the biggest variance are ploted in 3D, 2D and
1D.
}
\examples{
df <-   data.frame(S0 = c(900, 800, 500), # this is a data frame of 3 lines
                  I0 = c(100, 200, 500),
                  R0 = 0,
                  beta = c(1.4, 1.5, 1.6),
                  gamma = .15,
                  S = c(1, 2, 3),
                  I = c(2, 4, 6),
                  R = c(10, 20, 30),
                  nbiter = 1000,
                  seed = 123456789)
exp0 <- as_experiment(df, parameters = c("S0", "I0", "R0", "beta", "gamma"),
                      obsrates  = c("S", "I", "R"),
                      tmax = "nbiter", seed = "seed", experiment = "sir",
                      model =
                       system.file("models", "sir.gaml", package = "gamar"))
plot_params(exp0)
}
