% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repl.R
\name{repl}
\alias{repl}
\alias{repl.default}
\alias{repl.experiment}
\title{Replicate an experiment}
\usage{
repl(exp, times = NULL)

\method{repl}{default}(exp, times = NULL)

\method{repl}{experiment}(exp, times = NULL)
}
\arguments{
\item{exp}{An object of class \code{experiment}.}

\item{times}{Number of replicates or a named vector of rows to propagate in
\code{exp}.}
}
\value{
An object of class \code{experiment} that is the replication of the
inputed \code{x} argument.
}
\description{
Produces an \code{experiment} object from the replication of an
\code{experiment} object.
}
\details{
Parameters \code{times} and \code{exp} are exclusive.
If both specified, \code{times} will be taken.
}
\examples{
exp <- load_experiment("sir",
                       system.file("models", "sir.gaml", package = "gamar"))
exp1 <- repl(exp, times = 4)
exp1$seed <- c(1:4)
repl(exp1, times = c("1" = 3, "4" = 5))
}
