% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_to_gama.R
\name{save_to_gama}
\alias{save_to_gama}
\alias{save_to_gama.default}
\alias{save_to_gama.experiment}
\title{Save an experiment plan to a GAMA XML file}
\usage{
save_to_gama(exp, filename = NULL, path = NULL)

\method{save_to_gama}{default}(exp, filename, path)

\method{save_to_gama}{experiment}(exp, filename = NULL, path = NULL)
}
\arguments{
\item{exp}{An object of class \code{experiment}.}

\item{filename}{name of XML parameter file. If not
specified, name of `exp`` is used.}

\item{path}{Path to save `filename`. If not specified, current working
directory is used.}
}
\description{
Save an object of class \code{experiment} to an XML file GAMA-compliant.
}
\examples{
gaml_file <- system.file("models", "sir.gaml", package = "gamar")
exp1 <- load_experiment("sir", gaml_file)
save_to_gama(exp1)
}
