% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{setup_gama}
\alias{setup_gama}
\title{Download GAMA and configure}
\usage{
setup_gama(path = NA)
}
\arguments{
\item{path}{path to the applicatation Gama Platform}
}
\description{
Prompt the user to download GAMA if he wish to and then configures
\code{gamar}, linking it to a GAMA engine installed on the system.
}
\examples{
\dontrun{
# For interactive usage:
setup_gama()

# If input a path, the function will not be interactive:
setup_gama(getwd())
}

}
