% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sir1}
\alias{sir1}
\title{SIR epidemiological model}
\format{An object of class \code{experiment} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 1 rows and 11 columns.}
\usage{
sir1
}
\description{
An \code{experiment} of the SIR model with 1 simulation, 5 parameters and
3 monitored variables.
}
\keyword{datasets}
