% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{defpath}
\alias{defpath}
\title{Configure GAMA path and Java heap size}
\usage{
defpath(path, Xmx = "2048m", Xms = "512m")
}
\arguments{
\item{path}{Path to Gama}

\item{Xmx}{Maximum heap size}

\item{Xms}{Initial heap size}
}
\description{
Configure GAMA path and Java heap size
}
\examples{
\dontrun{
defpath(path = "/Applications/Gama.app/", Xmx = "4096m", Xms = "512m")
}
}
